[#175848172] Moved integration tests in a separated module
